==> 04_Machine_learning_with_Big_Data/02_handling-missing-values.ipynb <==
from pyspark.sql import SQLContext
import pyspark

sc = pyspark.SparkContext()
sqlContext = SQLContext(sc)

df = sqlContext.read.load('file:///home/hadoop/Downloads/big-data-4/daily_weather.csv', 
                          format='com.databricks.spark.csv', 
                          header='true',inferSchema='true')

df.describe().toPandas().transpose()

df.describe(['air_temp_9am']).show()

This says that there are 1090 rows. The total number of rows in the DataFrame is 1095:

df.count()

This means that 5 of the rows in the air_temp_9am column are missing values.

removeAllDF = df.na.drop()

removeAllDF.describe(['air_temp_9am']).show()

removeAllDF.count()

from pyspark.sql.functions import avg

imputeDF = df

removeAllDF.agg(avg('air_temp_9am')).first(), removeAllDF.agg(avg('air_temp_9am')).first()[0]

Iterating through each column in the DataFrame: we compute the mean value for that column and then replace any missing values in that column with the mean.

for x in imputeDF.columns:
    meanValue = removeAllDF.agg(avg(x)).first()[0]
    print(x,':', meanValue)
    imputeDF = imputeDF.na.fill(meanValue, [x])

df.describe(['air_temp_9am']).show()
imputeDF.describe(['air_temp_9am']).show()

imputeDF.count()

1 and 1, 1 & 1

If we remove all missing values from the data, how many air pressure at 9am measurements have values between 911.736 and 914.67?

removeAllDF.filter((removeAllDF.air_pressure_9am > 911.736) & (removeAllDF.air_pressure_9am < 914.67)).count()

If we impute the missing values with the minimum value, how many air temperature at 9am measurements are less than 42.292?

from pyspark.sql.functions import min

removeAllDF.select(min('air_temp_9am')).show()
removeAllDF.agg(min('air_temp_9am')).first()[0]

imputeDF_min = df

for x in imputeDF_min.columns:
    minValue = removeAllDF.agg(min(x)).first()[0]
    print(x,':', minValue)
    imputeDF_min = imputeDF_min.na.fill(minValue, [x])

imputeDF_min.filter((imputeDF_min.air_temp_9am < 42.292)).count()
