==> 04_Machine_learning_with_Big_Data/03_classification.ipynb <==
import pyspark
from pyspark.sql import SQLContext
from pyspark.sql import DataFrameNaFunctions
from pyspark.ml import Pipeline
from pyspark.ml.classification import DecisionTreeClassifier
from pyspark.ml.feature import Binarizer
from pyspark.ml.feature import VectorAssembler, StringIndexer, VectorIndexer

sc = pyspark.SparkContext()

sqlContext = SQLContext(sc)
df = sqlContext.read.load('file:///home/hadoop/Downloads/big-data-4/daily_weather.csv', 
                          format='com.databricks.spark.csv', 
                          header='true',inferSchema='true')
df.columns

 columns in the weather data we will use for the decision tree classifier

featureColumns = ['air_pressure_9am','air_temp_9am','avg_wind_direction_9am','avg_wind_speed_9am',
        'max_wind_direction_9am','max_wind_speed_9am','rain_accumulation_9am',
        'rain_duration_9am']

df = df.drop('number')

df = df.na.drop() 

df.count(), len(df.columns)

**Create categorical variable.** We create a categorical variable to denote if the humidity is not low. If the value is less than 25%, then we want the categorical value to be 0, otherwise the categorical value should be 1. We create this categorical variable as a column in a DataFrame using Binarizer:

binarizer = Binarizer(threshold=24.99999, inputCol="relative_humidity_3pm", outputCol="label")
binarizedDF = binarizer.transform(df)

The `threshold` argument specifies the threshold value for the variable, `inputCol` is the input column to read, and `outputCol` is the name of the new categorical column. The second line applies the `Binarizer` and creates a new DataFrame with the categorical column. We can look at the first four values in the new DataFrame:

binarizedDF.select("relative_humidity_3pm","label").show()

The first row's humidity value is greater than 25% and the label is 1. The other humidity values are less than 25% and have labels equal to 0.

**Aggregate features.** Let's aggregate the features we will use to make predictions into a single column:

assembler = VectorAssembler(inputCols=featureColumns, outputCol="features")
assembled = assembler.transform(binarizedDF)

The `inputCols` argument specifies our list of column names we defined earlier, and `outputCol` is the name of the new column. The second line creates a new DataFrame with the aggregated features in a column.

assembled.show(1)

assembled.toPandas()[:2]

Split training and test data by calling `randomSplit()`:

(trainingData, testData) = assembled.randomSplit([0.8,0.2], seed = 13234 )

trainingData.count(), testData.count()

DecisionTreeClassifier

dt = DecisionTreeClassifier(labelCol="label", featuresCol="features", maxDepth=5,
                            minInstancesPerNode=20, impurity="gini")

The `labelCol`argument is the column we are trying to predict, \
`featuresCol` specifies the aggregated features column, \
`maxDepth` is stopping criterion for tree induction based on maximum depth of tree, \
`minInstancesPerNode` is stopping criterion for tree induction based on minimum number of samples in a node, and \
`impurity` is the impurity measure used to split nodes.

pipeline = Pipeline(stages=[dt])
model = pipeline.fit(trainingData)

predictions = model.transform(testData)

predictions.select("prediction", "label").show(20)

save the predictions to a CSV file

predictions.select("prediction", "label")\
            .write\
            .save(path="file:///home/hadoop/big-data-processing/04_Machine_learning_with_Big_Data/predictions.csv",
                                                     format="com.databricks.spark.csv",
                                                     header='true')
