==> 04_Machine_learning_with_Big_Data/04_model-evaluation.ipynb <==
import pyspark
from pyspark.sql import SQLContext
from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.mllib.evaluation import MulticlassMetrics

sc = pyspark.SparkContext()

sqlContext = SQLContext(sc)

# predictions = sqlContext\
#             .read.load('file:///home/hadoop/big-data-processing/04_Machine_learning_with_Big_Data/predictions/part-00000-a44b280d-5e28-42cf-bd15-35306406af82-c000.csv', 
#                           format='com.databricks.spark.csv', 
#                           header='true',inferSchema='true')
# predictions = sqlContext\
#             .read.load('file:///home/hadoop/big-data-processing/04_Machine_learning_with_Big_Data/predictions.csv/*.csv', 
#                           format='com.databricks.spark.csv', 
#                           header='true',inferSchema='true')
predictions = sqlContext\
            .read.load('file:///home/hadoop/big-data-processing/04_Machine_learning_with_Big_Data/predictions.csv', 
                          format='com.databricks.spark.csv', 
                          header='true',inferSchema='true')

predictions.show(5)

create an instance of MulticlassClassificationEvaluator to determine the accuracy of the predictions:

https://spark.apache.org/docs/latest/api/python/reference/api/pyspark.ml.evaluation.MulticlassClassificationEvaluator.html

evaluator = MulticlassClassificationEvaluator(
    labelCol="label", predictionCol="prediction", metricName="precisionByLabel")

The first two arguments specify the names of the label and prediction columns, and the third argument specifies that we want the overall precision.

precision = evaluator.evaluate(predictions)
print("Precision = %g " % (precision))

evaluator2 = MulticlassClassificationEvaluator(
    labelCol="label", predictionCol="prediction", metricName="accuracy")
Accuracy = evaluator2.evaluate(predictions)
print("Accuracy = %g " % (Accuracy))

print ("Accuracy = %.4g" % (Accuracy * 100))

print ("Error = %g " % (1.0 - Accuracy))

**Display confusion matrix.** \
The MulticlassMetrics class can be used to generate a confusion matrix of our classifier model. However, unlike MulticlassClassificationEvaluator, MulticlassMetrics works with RDDs of numbers and not DataFrames, so we need to convert our predictions DataFrame into an RDD.

If we use the rdd attribute of predictions, we see this is an RDD of Rows:

predictions.rdd.take(2)

Instead, we can map the RDD to tuple to get an RDD of numbers:

predictions.rdd.map(tuple).take(2)

metrics = MulticlassMetrics(predictions.rdd.map(tuple))

metrics.confusionMatrix().toArray().transpose()

metrics.confusionMatrix().toArray()
